Added PowerUpsChecker and divied into packages
