+++ core/src/com/snake/game/gamescreens/MenuScreen.java
@@ -1,5 +1,5 @@
 
-package com.snake.game;
+package com.snake.game.gamescreens;
 
 
 import com.badlogic.gdx.Gdx;
@@ -12,6 +12,7 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.Touchable;
 import com.badlogic.gdx.scenes.scene2d.ui.Image;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
+import com.snake.game.SnakeGame;
 
 /**
  * Inital menu screen
