+++ core/src/com/snake/game/interfaces/Collectable.java
@@ -1,0 +1,5 @@
+package com.snake.game.interfaces;
+
+public abstract class Collectable implements IGameObject {
+    public abstract void moveToRandomPosition();
+}
