+++ core/src/com/snake/game/interfaces/IGameObject.java
@@ -1,0 +1,7 @@
+package com.snake.game.interfaces;
+
+import com.badlogic.gdx.math.Vector2;
+
+public interface IGameObject {
+    public Vector2 getPosition();
+}
