+++ core/src/com/snake/game/interfaces/IMovable.java
@@ -1,13 +1,8 @@
-package com.snake.game;
-
-import com.badlogic.gdx.math.Vector2;
+package com.snake.game.interfaces;
 
 /**
  * All object able to move should implement this interface
  */
 public interface IMovable {
-    public Vector2 getPosition();
-
     public void move(Movement movement);
-
 }
