+++ core/src/com/snake/game/interfaces/Movement.java
@@ -1,4 +1,4 @@
-package com.snake.game;
+package com.snake.game.interfaces;
 
 /**
  * Selects movement type for Snakes
