+++ core/src/com/snake/game/utilities/Random.java
@@ -1,4 +1,4 @@
-package com.snake.game;
+package com.snake.game.utilities;
 
 import java.util.concurrent.ThreadLocalRandom;
